+++ core/js/service-worker.php
@@ -7,9 +7,10 @@
  * @author     Dennis Suitters <dennis@diemen.design>
  * @copyright  2014-2019 Diemen Design
  * @license    http://opensource.org/licenses/MIT  MIT License
- * @version    0.0.8
+ * @version    0.0.9
  * @link       https://github.com/DiemenDesign/AuroraCMS
  * @notes      This PHP Script is designed to be executed using PHP 7+
+ * @changes    v0.0.9 Fix referencing Logo
  */
 header('Content-Type: application/javascript');
 header('Service-Worker-Allowed:/');
@@ -20,6 +21,19 @@ if(!defined('DS'))define('DS',DIRECTORY_SEPARATOR);
 if(!defined('THEME'))define('THEME','..'.DS.'..'.DS.'layout'.DS.$config['theme']);
 if(!defined('URL'))define('URL',PROTOCOL.$_SERVER['HTTP_HOST'].$settings['system']['url'].'/');
 $html=file_get_contents(THEME.DS."offline.html");
+$logo=THEME.DS.'images'.DS.'offlinelogo.png';
+if(file_exists(THEME.DS.'images'.DS.'offlinelogo.jpg'))
+  $logo=THEME.DS.'images'.DS.'offlinelogo.jpg';
+elseif(file_exists(THEME.DS.'images'.DS.'logo.png'))
+  $logo=THEME.DS.'images'.DS.'logo.png';
+elseif(file_exists(THEME.DS.'images'.DS.'logo.jpg'))
+  $logo=THEME.DS.'images'.DS.'logo.jpg';
+elseif(file_exists(THEME.DS.'images'.DS.str_replace(' ','-',$config['business']).'.png'))
+  $logo=THEME.DS.'images'.DS.str_replace(' ','-',$config['business']).'.png';
+elseif(file_exists(THEME.DS.'images'.DS.str_replace(' ','-',$config['business']).'.jpg'))
+  $logo=THEME.DS.'images'.DS.str_replace(' ','-',$config['business']).'.jpg';
+else
+  $logo=THEME.DS.'images'.DS.'offlinelogo.png';
 $html=preg_replace([
   '/<print seo=[\'\"]?title[\'\"]?>/',
   '/<print css=[\'\"]?bootstrap[\'\"]?>/',
@@ -34,9 +48,7 @@ $html=preg_replace([
   ($config['business']!=''?$config['business']:'AuroraCMS Offline'),
   file_get_contents(THEME.DS.'css'.DS.'bootstrap.min.css'),
   file_get_contents(THEME.DS.'css'.DS.'offline.css'),
-  (file_exists(THEME.DS.'images'.DS.'offlinelogo.png')?
-    'data:'.mime_content_type(THEME.DS.'images'.DS.'offlinelogo.png').';base64,'.base64_encode(file_get_contents(THEME.DS.'images'.DS.'offlinelogo.png')):
-    'data:'.mime_content_type(THEME.DS.'images'.DS.'logo.png').';base64,'.base64_encode(file_get_contents(THEME.DS.'images'.DS.'logo.png'))),
+  'data:'.mime_content_type($logo).';base64,'.base64_encode(file_get_contents($logo)),
   ($config['business']!=''?$config['business']:'AuroraCMS Offline Logo'),
   ($config['email']!=''?'Email: <a href="mailto:'.$config['email'].'">'.$config['email'].'</a>':''),
   ($config['phone']!=''?'Phone: <a href="tel:'.$config['phone'].'">'.$config['phone'].'</a>':''),
@@ -56,6 +68,7 @@ self.addEventListener("install",function(event){
       return cache.addAll(offlineFallbackPages);
     })
   );
+  self.skipWaiting();
 });
 
 self.addEventListener('fetch',event=>{