+++ core/layout/search.php
@@ -24,7 +24,7 @@ $search=isset($_POST['s'])?$_POST['s']:'%';?>
         <div class="form-group">
           <form method="post" action="<?php echo URL.$settings['system']['admin'].'/search';?>">
             <div class="input-group">
-              <input type="text" class="form-control" value="<?php echo$search;?>" name="s" placeholder="What are you looking for?">
+              <input type="text" class="form-control" value="<?php echo str_replace('%',' ',$search);?>" name="s" placeholder="What are you looking for?">
               <div class="input-group-append">
                 <button type="submit" class="btn btn-secondary">Go</button>
               </div>
@@ -34,8 +34,13 @@ $search=isset($_POST['s'])?$_POST['s']:'%';?>
       </div>
     </div>
     <div class="card bg-white pl-5">
-      <div class="card-body pl-5">
-<?php $s=$db->prepare("SELECT * FROM `".$prefix."content` WHERE LOWER(login_user) LIKE LOWER(:search) OR LOWER(contentType) LIKE LOWER(:search) OR LOWER(seoKeywords) LIKE LOWER(:search) OR LOWER(barcode) LIKE LOWER(:search) OR LOWER(fccid) LIKE LOWER(:search) OR LOWER(code) LIKE LOWER(:search) OR LOWER(brand) LIKE LOWER(:search) OR LOWER(title) LIKE LOWER(:search) OR LOWER(category_1) LIKE LOWER(:search) OR LOWER(category_2) LIKE LOWER(:search) OR LOWER(category_3) LIKE LOWER(:search) OR LOWER(category_4) LIKE LOWER(:search) OR LOWER(name) LIKE LOWER(:search) OR LOWER(email) LIKE LOWER(:search) OR LOWER(business) LIKE LOWER(:search) OR LOWER(address) LIKE LOWER(:search) OR LOWER(suburb) LIKE LOWER(:search) OR LOWER(city) LIKE LOWER(:search) OR LOWER(state) LIKE LOWER(:search) OR LOWER(postcode) LIKE LOWER(:search) OR LOWER(phone) LIKE LOWER(:search) OR LOWER(mobile) LIKE LOWER(:search) OR LOWER(fileALT) LIKE LOWER(:search) OR LOWER(attributionImageTitle) LIKE LOWER(:search) OR LOWER(attributionImageName) LIKE LOWER(:search) OR LOWER(cost) LIKE LOWER(:search) OR LOWER(subject) LIKE LOWER(:search) OR LOWER(notes) LIKE LOWER(:search) OR LOWER(attributionContentName) LIKE LOWER(:search) OR LOWER(tags) LIKE LOWER(:search) OR LOWER(seoTitle) LIKE LOWER(:search) OR LOWER(seoCaption) LIKE LOWER(:search) OR LOWER(seoDescription) LIKE LOWER(:search) ORDER BY ti DESC");
+      <div class="card-body p-5">
+<?php if($search=='modern day wizardry'){?>
+        <p class="text-dark">
+          The "Modern Day Wizard", Jeff Minter (<a href="http://www.minotaurproject.co.uk/jeff.php">Llamasoft</a>) has been the inspiration for me to develop software.
+        </p>
+<?php }else{
+$s=$db->prepare("SELECT * FROM `".$prefix."content` WHERE LOWER(login_user) LIKE LOWER(:search) OR LOWER(contentType) LIKE LOWER(:search) OR LOWER(seoKeywords) LIKE LOWER(:search) OR LOWER(barcode) LIKE LOWER(:search) OR LOWER(fccid) LIKE LOWER(:search) OR LOWER(code) LIKE LOWER(:search) OR LOWER(brand) LIKE LOWER(:search) OR LOWER(title) LIKE LOWER(:search) OR LOWER(category_1) LIKE LOWER(:search) OR LOWER(category_2) LIKE LOWER(:search) OR LOWER(category_3) LIKE LOWER(:search) OR LOWER(category_4) LIKE LOWER(:search) OR LOWER(name) LIKE LOWER(:search) OR LOWER(email) LIKE LOWER(:search) OR LOWER(business) LIKE LOWER(:search) OR LOWER(address) LIKE LOWER(:search) OR LOWER(suburb) LIKE LOWER(:search) OR LOWER(city) LIKE LOWER(:search) OR LOWER(state) LIKE LOWER(:search) OR LOWER(postcode) LIKE LOWER(:search) OR LOWER(phone) LIKE LOWER(:search) OR LOWER(mobile) LIKE LOWER(:search) OR LOWER(fileALT) LIKE LOWER(:search) OR LOWER(attributionImageTitle) LIKE LOWER(:search) OR LOWER(attributionImageName) LIKE LOWER(:search) OR LOWER(cost) LIKE LOWER(:search) OR LOWER(subject) LIKE LOWER(:search) OR LOWER(notes) LIKE LOWER(:search) OR LOWER(attributionContentName) LIKE LOWER(:search) OR LOWER(tags) LIKE LOWER(:search) OR LOWER(seoTitle) LIKE LOWER(:search) OR LOWER(seoCaption) LIKE LOWER(:search) OR LOWER(seoDescription) LIKE LOWER(:search) ORDER BY ti DESC");
 $s->execute([':search'=>'%'.$search.'%']);
 while($r=$s->fetch(PDO::FETCH_ASSOC)){?>
         <div id="google-title" class="contentType mt-4" data-contentType="<?php echo$r['contentType'];?>">
@@ -66,7 +71,8 @@ while($r=$s->fetch(PDO::FETCH_ASSOC)){?>
           else
             echo$config['seoDescription'];?>
         </div>
-<?php }?>
+<?php }
+}?>
       </div>
     </div>
   </div>