+++ core/layout/set_orders.php
@@ -7,12 +7,13 @@
  * @author     Dennis Suitters <dennis@diemen.design>
  * @copyright  2014-2019 Diemen Design
  * @license    http://opensource.org/licenses/MIT  MIT License
- * @version    0.0.8
+ * @version    0.0.9
  * @link       https://github.com/DiemenDesign/AuroraCMS
  * @notes      This PHP Script is designed to be executed using PHP 7+
  * @changes    v0.0.4 Fix Tooltips.
  * @changes    v0.0.7 Fix Width Formatting for better responsiveness.
  * @changes    v0.0.8 Add PayPal Client ID and Secret.
+ * @changes    v0.0.9 Add Option Toggle to Display Payment Options.
  */?>
 <main id="content" class="main">
   <ol class="breadcrumb">
@@ -59,6 +60,14 @@ while($rs=$ss->fetch(PDO::FETCH_ASSOC)){?>
             </div>
           </div>
 <?php }?>
+        </div>
+        <hr>
+        <legend>Payment Options</legend>
+        <div class="form-group row">
+          <div class="input-group col-4 col-sm-3 col-md-2 col-lg-3 col-xl-2">
+            <label class="switch switch-label switch-success"><input type="checkbox" id="options7" class="switch-input" data-dbid="1" data-dbt="config" data-dbc="options" data-dbb="7"<?php echo$config['options']{7}==1?' checked aria-checked="true"':' aria-checked="false"';?>><span class="switch-slider" data-checked="on" data-unchecked="off"></span></label>
+          </div>
+          <label for="options7" class="col-form-label col-8 col-sm-9 col-md-10 col-lg-9 col-xl-10">Display Payment Options Logo's (Logo's should be contained within the Tempate tags).</label>
         </div>
         <hr>
         <legend>Banking</legend>